move utils out of commands folder and fix clippy etc
