+++ cli/src/utils.rs
@@ -34,7 +34,7 @@ impl FileOrUrl {
                 Ok(bytes)
             }
             // Fetch from --url
-            (None, Some(uri)) => Ok(subxt_codegen::utils::fetch_metadata_bytes(&uri).await?),
+            (None, Some(uri)) => Ok(subxt_codegen::utils::fetch_metadata_bytes(uri).await?),
             // Default if neither is provided; fetch from local url
             (None, None) => {
                 let uri = "http://localhost:9933"
